Refactor scaling factor into shared utilities
